Ingame Wand GUI. Mostly untested (should be foolproof though, has lots of checks for nonexistent values)
GUI /should/ only display the essences that the current wand can handle, untested though.

Modified a couple ModularWand methods to be static, since they took an itemstack and modified it, without using any of the classes existing data. (They were required to calculate some values for the GUI)

Modified the texture, there was an extra pixel space and it was getting on my nerves